+++ world/examples/heightmap_visualization.rs
@@ -6,7 +6,7 @@ use rayon::ThreadPoolBuilder;
 use std::{fs::File, io::Write};
 use vek::*;
 use veloren_world::{
-    sim::{FileOpts, WorldOpts, DEFAULT_WORLD_MAP},
+    sim::{FileOpts, WorldOpts, DEFAULT_WORLD_MAP, DEFAULT_WORLD_SEED},
     Land, World,
 };
 
@@ -120,7 +120,7 @@ fn main() {
     let pool = ThreadPoolBuilder::new().build().unwrap();
     println!("Loading world");
     let (world, _index) = World::generate(
-        59686,
+        DEFAULT_WORLD_SEED,
         WorldOpts {
             seed_elements: true,
             world_file: FileOpts::LoadAsset(DEFAULT_WORLD_MAP.into()),